drivers: Rename imx6_net -> fec
