+++ src/drivers/net/fec/Mybuild
@@ -1,6 +1,6 @@
 package embox.driver.net
 
-module imx6_net {
+module fec {
 	option number log_level=0
 
 	option number base_addr=0x02188000
@@ -8,7 +8,7 @@ module imx6_net {
 
 	option number speed=100
 
-	source "imx6_net.c"
+	source "fec.c"
 
 	depends embox.net.skbuff
 	depends embox.net.l2.ethernet
