+++ templates/arm/imx6-wt/mods.config
@@ -23,7 +23,7 @@ configuration conf {
 	@Runlevel(1) include embox.driver.diag(impl="embox__driver__serial__imx_uart")
 	@Runlevel(0) include embox.driver.clock.cortexa9(periph_base_addr=0x00a00000,irq_num=29)
 
-	@Runlevel(2) include embox.driver.net.imx6_net(log_level=0, speed=1000)
+	@Runlevel(2) include embox.driver.net.fec(log_level=0, speed=1000)
 	@Runlevel(2) include embox.driver.net.loopback
 
 	include embox.fs.syslib.file_system_none
