+++ templates/arm/imx6/mods.config
@@ -49,7 +49,7 @@ configuration conf {
 	@Runlevel(2) include embox.net.skbuff(amount_skb=4000)
 	@Runlevel(2) include embox.net.skbuff_data(amount_skb_data=4000,data_size=1514,data_align=1,data_padto=1,ip_align=true)
 	@Runlevel(2) include embox.net.skbuff_extra(amount_skb_extra=128,extra_size=10,extra_align=1,extra_padto=1)
-	include embox.driver.net.imx6_net(log_level=0, speed=100)
+	include embox.driver.net.fec(log_level=0, speed=100)
 	include embox.driver.net.loopback
 
 	include embox.cmd.sys.env
